complete projects routes

+++ app/repositories/token_repositories/invitation_token_repository.py
@@ -18,6 +18,18 @@ async def get_invitation_token(
 
     return db_token
 
+async def get_user_invite_token(
+        db: AsyncSession,
+        user_id: int
+):
+    result = await db.execute(
+        select(ProjectInvitationTokens).where(
+            ProjectInvitationTokens.user_id == user_id
+        )
+    )
+
+    return result.scalars().first()
+
 async def save_invitation_token(
         db: AsyncSession,
         token_model: ProjectInvitationTokens
@@ -26,6 +38,33 @@ async def save_invitation_token(
     await db.flush()
     return token_model
 
-async def delete_invitation_token(db: AsyncSession, token: ProjectInvitationTokens) -> ProjectInvitationTokens | None:
-    await db.delete(token)
-    await db.commit()
+
+async def delete_invitation_token(
+        db: AsyncSession,
+        token_value: str  # Принимаем значение токена (строку)
+) -> ProjectInvitationTokens | None:
+    # Сначала находим токен по его значению
+    query = select(ProjectInvitationTokens).where(
+        ProjectInvitationTokens.hashed_token == token_value
+    )
+    result = await db.execute(query)
+    token = result.scalar_one_or_none()
+
+    if token:
+        await db.delete(token)
+        await db.commit()
+        return token
+
+    return None
+
+async def delete_invite_token_by_obj(
+        db: AsyncSession,
+        token_model: ProjectInvitationTokens
+) -> ProjectInvitationTokens | None:
+    try:
+        await db.delete(token_model)
+        await db.commit()
+        return token_model
+    except Exception as e:
+        await db.rollback()
+        raise e
